ticket-service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]==3.3.0
sqlalchemy>=1.4.0
httpx>=0.23.0